Add files via upload
This lab is a preparation for an upcoming group project on Kafka streams
. 
In this hands-on lab, I worked with Apache Kafka, used Docker to pull and run a local Kafka instance, and established a secure connection to the Kafka broker. 
I tested the connection, producing and consuming messages using Python. Then explore Kafka CLI, such as kcat, for inspecting and monitoring